#10: move avatar form to profile template
