--- src/pages/components/profileModal/profileModal.css
@@ -1,31 +1,0 @@
-.profile-modal {
-  height: 157px;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-around;
-  align-items: center;
-
-  &__label {
-    width: 25%;
-    text-align: center;
-    cursor: pointer;
-  }
-
-  &__input {
-    display: none;
-
-    &-text {
-      color: var(--ui-primary);
-      text-decoration: underline;
-
-      &_file {
-        color: var(--text-placeholder);
-        text-decoration: none;
-      }
-    }
-  }
-
-  &__error {
-    color: var(--danger);
-  }
-}
