--- src/pages/components/profileModal/types.ts
@@ -1,4 +1,0 @@
-export type ProfileModalProps = {
-  isOpen: boolean;
-  onClose: () => void;
-};
