+++ src/pages/components/profilePageTemplate/profilePageTemplate.css
@@ -13,3 +13,35 @@
     row-gap: 22px;
   }
 }
+
+.profile-modal {
+  height: 157px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-around;
+  align-items: center;
+
+  &__label {
+    width: 25%;
+    text-align: center;
+    cursor: pointer;
+  }
+
+  &__input {
+    display: none;
+
+    &-text {
+      color: var(--ui-primary);
+      text-decoration: underline;
+
+      &_file {
+        color: var(--text-placeholder);
+        text-decoration: none;
+      }
+    }
+  }
+
+  &__error {
+    color: var(--danger);
+  }
+}
